+++ frontend/src/components/index.ts
@@ -1,5 +1,5 @@
-export { default as FloatingBtn } from './FloatingBtn';
-export { default as Header } from './MainHeader';
-export { default as KioskLayout } from './KioskLayout';
-export { default as KeypadLayout } from './KeypadLayout';
-export { SCENE1_STEP_NAME } from './constData';
+export { default as FloatingBtn } from './common/FloatingBtn';
+export { default as Header } from './common/MainHeader';
+export { default as KioskLayout } from './common/KioskLayout';
+export { default as KeypadLayout } from './common/KeypadLayout';
+export { SCENE1_STEP_NAME } from '../consts/constData';
