+++ frontend/src/consts/index.ts
@@ -1,0 +1,1 @@
+export { MAIN_CONTENT_LIST, SCENE4_STEP_NAME, SCENE1_STEP_NAME } from './constData';
