+++ frontend/src/pages/index.ts
@@ -1,0 +1,2 @@
+export { default as MainPage } from './MainPage';
+export { default as Registered } from './Registered';
