+++ frontend/src/styles/registered/RegisteredStyle.tsx
@@ -1,0 +1,23 @@
+import styled from 'styled-components';
+
+export const Wrapper = styled.div`
+  height: 800px;
+  overflow: hidden;
+  position: relative;
+`;
+
+export const Container = styled.div`
+  display: flex;
+  flex-direction: column;
+  background: linear-gradient(to bottom, #878787, #15204c);
+  align-items: center;
+  height: 100%;
+`;
+
+export const Title = styled.div`
+  margin-top: 40px;
+  margin-bottom: 40px;
+  font-size: 24px;
+  text-align: left;
+  color: white;
+`;
